feat: create Pagination<T>
Pagination<T> is a generic record that encapsulates a pagination to be returned by the API. This record contains the current page, items per page, total items and a list of items.

And also a method called 'map', which accepts a function (Function) as an argument. This function is used to map list items of type T to another type R.